[SERVICE-508] Tidy for documentation output. Add environment validation
