+++ src/provider/utils/async.ts
@@ -11,3 +11,15 @@ export function deferredPromise<T = void>(): [Promise<T>, (value?: T) => void, (
     });
     return [p, res!, rej!];
 }
+
+/**
+ * Races a given promise against a timeout, and resolves to a `[didTimeout, value?]` tuple indicating
+ * whether the timeout occurred, and the value the promise resolved to (if timeout didn't occur)
+ * @param timeoutMs Timeout period in ms
+ * @param promise Promise to race against the timeout
+ */
+export function withTimeout<T>(timeoutMs: number, promise: Promise<T>): Promise<[boolean, T | undefined]> {
+    const timeout = new Promise<[boolean, undefined]>(res => setTimeout(() => res([true, undefined]), timeoutMs));
+    const p = promise.then(value => ([false, value] as [boolean, T]));
+    return Promise.race([timeout, p]);
+}
