+++ test/demo/utils/fdc3RemoteExecution.ts
@@ -13,8 +13,8 @@
 import {Identity} from 'openfin/_v2/main';
 import {WindowOption} from 'openfin/_v2/api/window/windowOption';
 
-import {Application, Context, IntentType, AppIntent, ChannelId, DefaultChannel, DesktopChannel, Channel, deserializeError, FDC3Event, FDC3EventType} from '../../../src/client/main';
-import {RaiseIntentPayload} from '../../../src/client/internal';
+import {Application, Context, IntentType, AppIntent, ChannelId, FDC3Event, FDC3EventType} from '../../../src/client/main';
+import {RaiseIntentPayload, deserializeError} from '../../../src/client/internal';
 
 import {OFPuppeteerBrowser, TestWindowContext, TestChannelTransport} from './ofPuppeteer';
 import {RemoteChannel} from './RemoteChannel';
